+++ api/controllers/AuthController.js
@@ -4,45 +4,13 @@ const jwt = require('jsonwebtoken');
 
 const User = require('../models/Users');
 const v = require('../middleware/validators/AuthValidator');
+const q = require('../middleware/query-executors/AuthQueryExecutor');
 
 exports.jwtSignupPost = [
   ...v.validateSignupBody,
-  (req, res) => {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      //  maybe there should be a different error code
-      return res.status(404).send({ errors: errors.array() });
-    }
-    try {
-      bcrypt.hash(req.body.password, 10, (err, hashed) => {
-        if (err) {
-          throw (err);
-        }
-        const user = new User({
-          first_name: req.body.first_name,
-          last_name: req.body.last_name,
-          username: req.body.username,
-          password: hashed,
-          email: req.body.email,
-          dob: req.body.dob,
-          friends: [],
-          bio: req.body.bio,
-        });
-        user.save((error) => {
-          if (error) {
-            throw (error);
-          }
-          /**
-           * Can set up JWT token to be sent here on the sign up
-           * and can have it on sign in
-           */
-        });
-      });
-    } catch (error) {
-      return res.status(400).send({ errors: [{ msg: 'internal errorf' }] });
-    }
-    return res.status(200).send({ success: [{ msg: 'Thanks for signing up' }] });
-  },
+  q.saveNewUser(),
+  (req, res) => res.status(200).send({ success: [{ msg: 'Thanks for signing up' }] }),
+
 ];
 
 exports.jwtLoginPost = [
