--- api/middleware/responders/PostsResponder.js
@@ -1,8 +1,0 @@
-exports.sendFeedResponse = (req, res) => {};
-exports.sendGetPostResponse = (req, res) => {};
-exports.sendGetCommentsResponse = (req, res) => {};
-exports.sendGetLikesResponse = (req, res) => {};
-exports.sendCreatePostResponse = (req, res) => {};
-exports.sendLikePostResponse = (req, res) => {};
-exports.sendCommentPostResponse = (req, res) => {};
-exports.sendDeletePostResponse = (req, res) => {};
